Switch back to stable NBT
The current game plan is to roll out BedrockProtocol only with ext-encoding first. This ensures that
if there are any major issues that my (extensive) unit tests haven't caught, the damage will be
limited to packet encoding, so no permanent damage will be done.

Since we use NBT in world saves as well as on the network, depending on the version using
ext-encoding here would make this next upgrade very risky, so it's not worth it for now. The time
for that will come in the next month or two.