Migrate encode & decode to use ext-encoding (really big commit) (#322)
- PacketSerializer is now gone
- All places which previously used ByteBufferReader and ByteBufferWriter instead of PacketSerializer/BinaryStream
- All helper methods previously in PacketSerializer are now provided as CommonTypes:: static methods instead

BinaryUtils is still required for now since there's a couple of places where it's still irreplaceable. But all encoding and decoding is now done using ext-encoding, which should yield significant performance improvements.

Stable version of NBT is still used for the time being as we don't want to be forced to fully switch over to ext-encoding right away. Doing network only is safer for now since there'll be no lasting damage done if there are growing pains. Proper integration of ext-encoding+NBT will come later. For now, this is enough to net some gains.